refactor: Add Logout component button
